Move various filesystem-related facilities to a dedicated package
