+++ protobuf-maven-plugin/src/main/java/io/github/ascopes/protobufmavenplugin/dependencies/aether/AetherArtifactMapper.java
@@ -18,7 +18,7 @@ package io.github.ascopes.protobufmavenplugin.dependencies.aether;
 import static java.util.Objects.requireNonNullElse;
 import static java.util.function.Predicate.not;
 
-import io.github.ascopes.protobufmavenplugin.utils.FileUtils;
+import io.github.ascopes.protobufmavenplugin.fs.FileUtils;
 import java.nio.file.Path;
 import java.util.Collection;
 import java.util.Objects;
