+++ protobuf-maven-plugin/src/main/java/io/github/ascopes/protobufmavenplugin/fs/FileUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.ascopes.protobufmavenplugin.utils;
+package io.github.ascopes.protobufmavenplugin.fs;
 
 import java.io.IOException;
 import java.nio.file.FileSystem;
